Implement experimental support for shared-table multitenancy
The config adds a new `tenant` option which defaults to empty, keeping current behavior (and schemas) for existing users.
If a tenant is specified, the tables are expected to have a `tenant` column/partitioning, which is used to separate the data.
New tables are created with this column automatically IF a tenant name is specified, otherwise it's left out.

So there are two schema variants - the existing/default singletenant schema without the extra column, and a multitenant schema with the added column if a tenant name is provided in the config.

This could be mixed with the existing per-DB multitenant support depending on the context.

Signed-off-by: Nick Parker <nparker@gitlab.com>